Login & CRU Apostolat & Model Relations

+++ database/migrations/2022_01_22_103707_create_sous_zones_table.php
@@ -17,7 +17,7 @@ class CreateSousZonesTable extends Migration
             $table->id();
             $table->string('nom');
             $table->string('quartier');
-            $table->unsignedInteger('zone_id');
+            $table->unsignedBigInteger('zone_id');
             $table->timestamps();
             $table->softDeletes();
             $table->foreign('zone_id')->references('id')->on('zones');
